feat: Remove PDF support
The PDF support is now done by chat-ui with RAG.
The OpenAI API implementation should remain OpenAI API compatible, and
this should focus on serving model as API only.

If we wish to implement RAG or something, we need to develop a separated
RAG agent. (like what chat-ui is doing now).

We have a plan to move the RAG part of chat-ui to our own Python API
server.